requirements.txt
Flask==3.0.3
gunicorn
psycopg2-binary==2.9.9